Added weebly site's files
The files are there and the `splitTest.php` entries are updated, so the
split testing functionality works.  The forms need to be updated
according to the `templatehtml.php` file listed under the `_html`
directory and the file's resources need to be located in the file
structure's `_assets` and `_js` directories.